Richard Comment
The first part of Nav Headers has no visible effect. So it might be good to take them off. I added a css file called edit.css, which supports the "wrap" div I put to give some space around the three boxes. All changes will be in a new branch so that it will be easy for you to roll-back to the original version.